moved queries to separate file
